[step2] refactoring : 로또 기계가 랜덤값을 생성하도록 역할 분리
