+++ src/test/kotlin/lotto/LottoTest.kt
@@ -7,11 +7,11 @@ internal class LottoTest {
 
     @Test
     fun `로또는 6개의 숫자를 갖는다`() {
-        assertThat(Lotto().numbers.size).isEqualTo(6)
+        assertThat(Lotto(listOf(1, 2, 3, 4, 5, 6)).numbers.size).isEqualTo(6)
     }
 
     @Test
     fun `로또는 6개의 중복되지 않는 숫자를 갖는다`() {
-        assertThat(Lotto().numbers.distinct().size).isEqualTo(6)
+        assertThat(Lotto(listOf(1, 2, 3, 4, 5, 6)).numbers.distinct().size).isEqualTo(6)
     }
 }
